homework-7
