+++ main.js
@@ -1,228 +1,15 @@
-window.addEventListener('DOMContentLoaded', function () {
-  const $arenas = document.querySelector('.arenas');
-  const $formFight = document.querySelector('.control');
-  const $buttonFight = document.querySelector('.buttonWrap .button');
-  const $chat = document.querySelector('.chat');
-
-  const player1 = {
-    player: 1,
-    name: 'Kitana',
-    hp: 100,
-    img: 'http://reactmarathon-api.herokuapp.com/assets/kitana.gif',
-    weapon: [],
-    changeHp,
-    elHp,
-    renderHp
-  };
-  const player2 = {
-    player: 2,
-    name: 'Subzero',
-    hp: 100,
-    img: 'http://reactmarathon-api.herokuapp.com/assets/subzero.gif',
-    weapon: [],
-    changeHp,
-    elHp,
-    renderHp
-  };
-
-  const HIT = {
-    head: 30,
-    body: 25,
-    foot: 20,
- };
- const ATTACK = ['head', 'body', 'foot'];
- const logs = {
-  start: 'Часы показывали [time], когда [player1] и [player2] бросили вызов друг другу.',
-  end: [
-      'Результат удара [playerWins]: [playerLose] - труп',
-      '[playerLose] погиб от удара бойца [playerWins]',
-      'Результат боя: [playerLose] - жертва, [playerWins] - убийца',
-  ],
-  hit: [
-      '[playerDefence] пытался сконцентрироваться, но [playerKick] разбежавшись раздробил копчиком левое ухо врага.',
-      '[playerDefence] расстроился, как вдруг, неожиданно [playerKick] случайно раздробил грудью грудину противника.',
-      '[playerDefence] зажмурился, а в это время [playerKick], прослезившись, раздробил кулаком пах оппонента.',
-      '[playerDefence] чесал <вырезано цензурой>, и внезапно неустрашимый [playerKick] отчаянно размозжил грудью левый бицепс оппонента.',
-      '[playerDefence] задумался, но внезапно [playerKick] случайно влепил грубый удар копчиком в пояс оппонента.',
-      '[playerDefence] ковырялся в зубах, но [playerKick] проснувшись влепил тяжелый удар пальцем в кадык врага.',
-      '[playerDefence] вспомнил что-то важное, но внезапно [playerKick] зевнув, размозжил открытой ладонью челюсть противника.',
-      '[playerDefence] осмотрелся, и в это время [playerKick] мимоходом раздробил стопой аппендикс соперника.',
-      '[playerDefence] кашлянул, но внезапно [playerKick] показав палец, размозжил пальцем грудь соперника.',
-      '[playerDefence] пытался что-то сказать, а жестокий [playerKick] проснувшись размозжил копчиком левую ногу противника.',
-      '[playerDefence] забылся, как внезапно безумный [playerKick] со скуки, влепил удар коленом в левый бок соперника.',
-      '[playerDefence] поперхнулся, а за это [playerKick] мимоходом раздробил коленом висок врага.',
-      '[playerDefence] расстроился, а в это время наглый [playerKick] пошатнувшись размозжил копчиком губы оппонента.',
-      '[playerDefence] осмотрелся, но внезапно [playerKick] робко размозжил коленом левый глаз противника.',
-      '[playerDefence] осмотрелся, а [playerKick] вломил дробящий удар плечом, пробив блок, куда обычно не бьют оппонента.',
-      '[playerDefence] ковырялся в зубах, как вдруг, неожиданно [playerKick] отчаянно размозжил плечом мышцы пресса оппонента.',
-      '[playerDefence] пришел в себя, и в это время [playerKick] провел разбивающий удар кистью руки, пробив блок, в голень противника.',
-      '[playerDefence] пошатнулся, а в это время [playerKick] хихикая влепил грубый удар открытой ладонью по бедрам врага.',
-  ],
-  defence: [
-      '[playerKick] потерял момент и храбрый [playerDefence] отпрыгнул от удара открытой ладонью в ключицу.',
-      '[playerKick] не контролировал ситуацию, и потому [playerDefence] поставил блок на удар пяткой в правую грудь.',
-      '[playerKick] потерял момент и [playerDefence] поставил блок на удар коленом по селезенке.',
-      '[playerKick] поскользнулся и задумчивый [playerDefence] поставил блок на тычок головой в бровь.',
-      '[playerKick] старался провести удар, но непобедимый [playerDefence] ушел в сторону от удара копчиком прямо в пятку.',
-      '[playerKick] обманулся и жестокий [playerDefence] блокировал удар стопой в солнечное сплетение.',
-      '[playerKick] не думал о бое, потому расстроенный [playerDefence] отпрыгнул от удара кулаком куда обычно не бьют.',
-      '[playerKick] обманулся и жестокий [playerDefence] блокировал удар стопой в солнечное сплетение.'
-  ],
-  draw: 'Ничья - это тоже победа!'
-};
+  import {player1, player2, enemyAttack, playerAttack} from './modules/players.js';
+  import { createPlayer, showResult } from './modules/elements.js';
+  import { generateLogs } from './modules/logs.js';
+  import { $arenas, $formFight } from './modules/storage.js';
+  
+ 
   generateLogs('start', player1, player2);
-  function createElement(tag, className){
-    $tag = document.createElement(tag);
-    if(className){
-      $tag.classList.add(className);
-    }
-   
-    return $tag;
-  }
-
-  function createPlayer(obj) {
-    const $player = createElement('div', `player${obj.player}`);
-    const $progressbar = createElement('div', 'progressbar');
-    const $life = createElement('div', 'life');
-    const $name = createElement('div', 'name');
-    const $character = createElement('div', 'character');
-    const $playerImg = createElement('img');
-
-    $life.style.width = `${obj.hp}%`;
-    $playerImg.src = obj.img;
-    $name.innerHTML = obj.name;
-
-    $progressbar.appendChild($life);
-    $progressbar.appendChild($name);
-    $player.appendChild($progressbar);
-    $character.appendChild($playerImg);
-    $player.appendChild($character);
-    
-    return $player;
-  };
 
   $arenas.appendChild(createPlayer(player1));
   $arenas.appendChild(createPlayer(player2));
 
-  function getRandom(n){
-    return Math.ceil(Math.random()*n);
-  }
-  
-  function insertChat(text){
-    let el = `<p>${text}</p>`;
-    $chat.insertAdjacentHTML('beforeend', el);
-  }
-  function generateLogs(type, player1, player2, player, value){
-    let date = new Date();
-    let timeNow = `${date.getHours()}:${date.getMinutes()}`;
-  
-    switch (type){
-      case 'start':
-        insertChat(logs[type].replace('[time]', timeNow).replace('[player1]', player1.name).replace('[player2]', player2.name)); 
-        break;
-      case 'hit': 
-        insertChat(`${timeNow} ${logs[type][getRandom(logs.hit.length)-1].replace('[playerKick]', player1.name).replace('[playerDefence]', player2.name)} -${value} [${player.hp}/100]`);
-        break;
-      case 'defence':
-        insertChat(`${timeNow} ${logs[type][getRandom(logs.defence.length)-1].replace('[playerKick]', player1.name).replace('[playerDefence]', player2.name)} `);
-        break;
-      case 'draw':
-        insertChat(`${timeNow} ${logs[type]}`);
-        break;
-      case 'end':
-        insertChat( `${timeNow} ${logs[type][getRandom(logs.end.length)-1].replace('[playerWins]', player1.name).replace('[playerLose]', player2.name)}`);
-        break;
-    }
-  }
-  
-  function showResultText(name){
-    const $winTitle = createElement('div', 'winTitle');
-    if(name){
-      $winTitle.innerHTML = name + ' win!!!';
-    }
-    else {
-      $winTitle.innerHTML = 'draw!';
-    }
-    $arenas.prepend($winTitle);
-   
-  }
-  
-  function createReloadButton(){
-    const $reloadWrap = createElement('div', 'reloadWrap');
-    const $reloadButton = createElement('button', 'button');
-    $reloadButton.innerHTML = 'Restart';
-    $reloadWrap.appendChild($reloadButton);
-    $arenas.appendChild($reloadWrap)
-    $reloadButton.addEventListener('click',function(){
-      window.location.reload();
-    } )
-  }
  
-
-  function changeHp(n){
-    this.hp-=n;
-    if(this.hp<=n){
-      this.hp = 0;
-    }
-
-    return this.hp;
-  }
-
-  function elHp(){
-    return document.querySelector(`.player${this.player} .life`);
-  }
-
-  function renderHp(){
-  
-      this.elHp().style.width = `${this.hp}%`;
-
-  }
-function enemyAttack(){
-  const hit = ATTACK[getRandom(3)-1];
-  const defence = ATTACK[getRandom(3)-1];
-  return {
-    value: getRandom(HIT[hit]),
-    hit,
-    defence
-  }
-}
- function playerAttack(){
-  const attack = {};
-
-  for (let item of $formFight){
-    if (item.checked && item.name === 'hit'){
-      attack.value = getRandom(HIT[item.value]);
-      attack.hit = item.value;
-    }
-    if (item.checked && item.name === 'defence'){
-      attack.defence = item.value
-    }
-    item.checked = false;
-  }
-  return attack;
- }
-
- function showResult(){
-  if (player1.hp === 0 || player2.hp === 0){
-    $buttonFight.style.backgroundColor = 'grey';
-    $buttonFight.disabled = true;
-   createReloadButton();
-  }
-  if (player1.hp === 0 && player1.hp < player2.hp){
-    showResultText(player2.name);
-    generateLogs('end', player2, player1);
-  }
-  else if (player2.hp === 0 && player2.hp < player1.hp){
-    showResultText(player1.name);
-    generateLogs('end', player1, player2);
-  }
-
-  else if (player2.hp === 0 && player2.hp === 0) {
-    showResultText();
-    generateLogs('draw');
-  }
- }
-
   $formFight.addEventListener('submit', function(e){
     e.preventDefault();
     const enemy = enemyAttack();
@@ -249,4 +36,3 @@ function enemyAttack(){
   })
 
 
-});
